+++ lib/composite_unit_measurements/length.rb
@@ -17,6 +17,7 @@ module CompositeUnitMeasurements
       # @param [String] string The string to parse for length measurement.
       # @return [UnitMeasurements::Length]
       #   Returns a UnitMeasurements::Length object if parsing is successful.
+      #
       # @raise [UnitMeasurements::ParseError]
       #   If the string does not match any known format.
       #
@@ -24,8 +25,10 @@ module CompositeUnitMeasurements
       # @since 0.2.0
       def parse(string)
         case string
-        when FOOT_INCH then parse_foot_inch(string)
-        else                raise UnitMeasurements::ParseError, string
+        when FOOT_INCH        then parse_foot_inch(string)
+        when KILOMETRE_METRE  then parse_kilometre_metre(string)
+        when METRE_CENTIMETRE then parse_metre_centimetre(string)
+        else                       raise UnitMeasurements::ParseError, string
         end
       end
 
@@ -46,7 +49,45 @@ module CompositeUnitMeasurements
         foot, inch = string.match(FOOT_INCH)&.captures
 
         if foot && inch
-          UnitMeasurements::Length.new(foot, :ft) + UnitMeasurements::Length.new(inch, :in)
+          UnitMeasurements::Length.new(foot, "ft") + UnitMeasurements::Length.new(inch, "in")
+        end
+      end
+
+      # @private
+      # Parses a +string+ representing a length in the format of +metre-centimetre+.
+      #
+      # @param [String] string
+      #   The string representing length measurement in the format of *metre-centimetre*.
+      # @return [UnitMeasurements::Length]
+      #   Returns a UnitMeasurements::Length object if parsing is successful.
+      #
+      # @see METRE_CENTIMETRE
+      # @author {Harshal V. Ladhe}[https://shivam091.github.io/]
+      # @since 0.3.0
+      def parse_metre_centimetre(string)
+        metre, centimetre = string.match(METRE_CENTIMETRE)&.captures
+
+        if metre && centimetre
+          UnitMeasurements::Length.new(metre, "m") + UnitMeasurements::Length.new(centimetre, "cm")
+        end
+      end
+
+      # @private
+      # Parses a +string+ representing a length in the format of +kilometre-metre+.
+      #
+      # @param [String] string
+      #   The string representing length measurement in the format of *kilometre-metre*.
+      # @return [UnitMeasurements::Length]
+      #   Returns a UnitMeasurements::Length object if parsing is successful.
+      #
+      # @see KILOMETRE_METRE
+      # @author {Harshal V. Ladhe}[https://shivam091.github.io/]
+      # @since 0.3.0
+      def parse_kilometre_metre(string)
+        kilometre, metre = string.match(KILOMETRE_METRE)&.captures
+
+        if kilometre && metre
+          UnitMeasurements::Length.new(kilometre, "km") + UnitMeasurements::Length.new(metre, "m")
         end
       end
     end
@@ -65,10 +106,40 @@ module CompositeUnitMeasurements
     # @since 0.2.0
     INCH_ALIASES = /(?:"|in|inch(?:es)?)/.freeze
 
+    # Regex pattern for aliases of +metre+ unit.
+    #
+    # @author {Harshal V. Ladhe}[https://shivam091.github.io/]
+    # @since 0.3.0
+    METRE_ALIASES = /(?:m|meter(?:s)?|metre(?:s)?)/.freeze
+
+    # Regex pattern for aliases of +centimetre+ unit.
+    #
+    # @author {Harshal V. Ladhe}[https://shivam091.github.io/]
+    # @since 0.3.0
+    CENTIMETRE_ALIASES = /(?:cm|centimeter(?:s)?|centimetre(?:s)?)/.freeze
+
+    # Regex pattern for aliases of +kilometre+ unit.
+    #
+    # @author {Harshal V. Ladhe}[https://shivam091.github.io/]
+    # @since 0.3.0
+    KILOMETRE_ALIASES = /(?:km|kilometer(?:s)?|kilometre(?:s)?)/.freeze
+
     # Regex pattern for parsing a length measurement in the format of +foot-inch+.
     #
     # @author {Harshal V. Ladhe}[https://shivam091.github.io/]
     # @since 0.2.0
     FOOT_INCH = /\A#{ANY_NUMBER}\s*#{FOOT_ALIASES}\s*#{ANY_NUMBER}\s*#{INCH_ALIASES}\z/.freeze
+
+    # Regex pattern for parsing a length measurement in the format of +metre-centimetre+.
+    #
+    # @author {Harshal V. Ladhe}[https://shivam091.github.io/]
+    # @since 0.3.0
+    METRE_CENTIMETRE = /\A#{ANY_NUMBER}\s*#{METRE_ALIASES}\s*#{ANY_NUMBER}\s*#{CENTIMETRE_ALIASES}\z/.freeze
+
+    # Regex pattern for parsing a length measurement in the format of +kilometre-metre+.
+    #
+    # @author {Harshal V. Ladhe}[https://shivam091.github.io/]
+    # @since 0.3.0
+    KILOMETRE_METRE = /\A#{ANY_NUMBER}\s*#{KILOMETRE_ALIASES}\s*#{ANY_NUMBER}\s*#{METRE_ALIASES}\z/.freeze
   end
 end