+++ lib/composite_unit_measurements/weight.rb
@@ -24,9 +24,10 @@ module CompositeUnitMeasurements
       # @since 0.2.0
       def parse(string)
         case string
-        when POUND_OUNCE then parse_pound_ounce(string)
-        when STONE_POUND then parse_stone_pound(string)
-        else                  raise UnitMeasurements::ParseError, string
+        when POUND_OUNCE       then parse_pound_ounce(string)
+        when STONE_POUND       then parse_stone_pound(string)
+        when KILOGRAMME_GRAMME then parse_kilogramme_gramme(string)
+        else                        raise UnitMeasurements::ParseError, string
         end
       end
 
@@ -47,7 +48,7 @@ module CompositeUnitMeasurements
         pound, ounce = string.match(POUND_OUNCE)&.captures
 
         if pound && ounce
-          UnitMeasurements::Weight.new(pound, :lb) + UnitMeasurements::Weight.new(ounce, :oz)
+          UnitMeasurements::Weight.new(pound, "lb") + UnitMeasurements::Weight.new(ounce, "oz")
         end
       end
 
@@ -66,7 +67,26 @@ module CompositeUnitMeasurements
         stone, pound = string.match(STONE_POUND)&.captures
 
         if stone && pound
-          UnitMeasurements::Weight.new(stone, :st) + UnitMeasurements::Weight.new(pound, :lb)
+          UnitMeasurements::Weight.new(stone, "st") + UnitMeasurements::Weight.new(pound, "lb")
+        end
+      end
+
+      # @private
+      # Parses a +string+ representing a weight in the format of +kilogramme-gramme+.
+      #
+      # @param [String] string
+      #   The string representing weight measurement in the format of *kilogramme-gramme*.
+      # @return [UnitMeasurements::Weight]
+      #   Returns a UnitMeasurements::Weight object if parsing is successful.
+      #
+      # @see KILOGRAMME_GRAMME
+      # @author {Harshal V. Ladhe}[https://shivam091.github.io/]
+      # @since 0.3.0
+      def parse_kilogramme_gramme(string)
+        kilogramme, gramme = string.match(KILOGRAMME_GRAMME)&.captures
+
+        if kilogramme && gramme
+          UnitMeasurements::Weight.new(kilogramme, "kg") + UnitMeasurements::Weight.new(gramme, "g")
         end
       end
     end
@@ -77,30 +97,48 @@ module CompositeUnitMeasurements
     #
     # @author {Harshal V. Ladhe}[https://shivam091.github.io/]
     # @since 0.2.0
-    POUND_UNITS = /(?:#|lb|lbs|lbm|pound-mass|pound(?:s)?)/.freeze
+    POUND_ALIASES = /(?:#|lb|lbs|lbm|pound-mass|pound(?:s)?)/.freeze
 
     # Regex pattern for aliases of +ounce+ unit.
     #
     # @author {Harshal V. Ladhe}[https://shivam091.github.io/]
     # @since 0.2.0
-    OUNCE_UNITS = /(?:oz|ounce(?:s)?)/.freeze
+    OUNCE_ALIASES = /(?:oz|ounce(?:s)?)/.freeze
 
     # Regex pattern for aliases of +stone+ unit.
     #
     # @author {Harshal V. Ladhe}[https://shivam091.github.io/]
     # @since 0.2.0
-    STONE_UNITS = /(?:st|stone(?:s)?)/.freeze
+    STONE_ALIASES = /(?:st|stone(?:s)?)/.freeze
+
+    # Regex pattern for aliases of +gramme+ unit.
+    #
+    # @author {Harshal V. Ladhe}[https://shivam091.github.io/]
+    # @since 0.3.0
+    GRAMME_ALIASES = /(?:g|gram(?:s)?|gramme(?:s)?)/.freeze
+
+    # Regex pattern for aliases of +kilogramme+ unit.
+    #
+    # @author {Harshal V. Ladhe}[https://shivam091.github.io/]
+    # @since 0.3.0
+    KILOGRAMME_ALIASES = /(?:kg|kilogram(?:s)?|kilogramme(?:s)?)/.freeze
 
     # Regex pattern for parsing a weight measurement in the format of +pound-ounce+.
     #
     # @author {Harshal V. Ladhe}[https://shivam091.github.io/]
     # @since 0.2.0
-    POUND_OUNCE = /\A#{ANY_NUMBER}\s*#{POUND_UNITS}\s*#{ANY_NUMBER}\s*#{OUNCE_UNITS}\z/.freeze
+    POUND_OUNCE = /\A#{ANY_NUMBER}\s*#{POUND_ALIASES}\s*#{ANY_NUMBER}\s*#{OUNCE_ALIASES}\z/.freeze
 
     # Regex pattern for parsing a weight measurement in the format of +stone-pound+.
     #
     # @author {Harshal V. Ladhe}[https://shivam091.github.io/]
     # @since 0.2.0
-    STONE_POUND = /\A#{ANY_NUMBER}\s*#{STONE_UNITS}\s*#{ANY_NUMBER}\s*#{POUND_UNITS}\z/.freeze
+    STONE_POUND = /\A#{ANY_NUMBER}\s*#{STONE_ALIASES}\s*#{ANY_NUMBER}\s*#{POUND_ALIASES}\z/.freeze
+
+    # Regex pattern for parsing a weight measurement in the format of +kilogramme-gramme+.
+    #
+    # @author {Harshal V. Ladhe}[https://shivam091.github.io/]
+    # @since 0.3.0
+    KILOGRAMME_GRAMME = /\A#{ANY_NUMBER}\s*#{KILOGRAMME_ALIASES}\s*#{ANY_NUMBER}\s*#{GRAMME_ALIASES}\z/.freeze
   end
 end